Performance, Maintenace and Bug Fixes

+++ Command/BulkProductSync.php
@@ -4,7 +4,7 @@
  *
  * @category   Camiloo
  * @package    Camiloo_Channelunity
- * @copyright  Copyright (c) 2019 ChannelUnity Limited (http://www.channelunity.com)
+ * @copyright  Copyright (c) 2016-2024 ChannelUnity Limited (http://www.channelunity.com)
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -17,7 +17,11 @@ use Camiloo\Channelunity\Model\Helper;
 use Camiloo\Channelunity\Model\Products;
 use Magento\Framework\App\ResourceConnection;
 use Magento\Framework\App\State;
-use \Magento\Store\Model\StoreManagerInterface;
+use Magento\Store\Model\StoreManagerInterface;
+use Magento\Framework\Console\Cli;
+use Magento\Framework\App\Area;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Store\Model\Store;
 
 /**
  * Sends product updates in bulk to ChannelUnity. This is used as a result
@@ -31,7 +35,7 @@ class BulkProductSync extends Command
     private $state;
     private $storeManager;
     private $syncLimit = 50;
-    
+
     public function __construct(
         Helper $helper,
         Products $cuproducts,
@@ -46,26 +50,28 @@ class BulkProductSync extends Command
         $this->state = $state;
         $this->storeManager = $storeManager;
     }
-    
+
     /**
      * {@inheritdoc}
      */
     protected function configure()
     {
         $this->setName('channelunity:sync_products_bulk')
-            ->setDescription("Sync products to ChannelUnity that have been recently saved in bulk ($this->syncLimit max per run)");
+            ->setDescription("Sync products to ChannelUnity that have been recently saved in bulk ({$this->syncLimit} max per run)");
         parent::configure();
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         // Manual run
         $this->updateBatch($output);
+
+        return Cli::RETURN_SUCCESS;
     }
-    
+
     //======================= Cron commands =================================//
 
     public function every5Minutes()
@@ -73,82 +79,86 @@ class BulkProductSync extends Command
         // Crontab run
         $this->updateBatch();
     }
-    
+
     //======================= Private methods =================================//
-    private function updateBatch($output = null)
+
+    private function updateBatch(OutputInterface $output = null)
     {
         try {
-            $this->state->setAreaCode("global");
-        } catch (\Exception $e) {
+            $this->state->setAreaCode(Area::AREA_GLOBAL);
+        } catch (LocalizedException $e) {
+            // Area code is already set, safely ignore
         }
+
         $startTime = time();
-        
+
         // If a bulk edit recently happened, we'll have stored some product IDs
         $productUpdatesTable = $this->resource->getTableName('product_updates');
-        $connection = $this->resource->getConnection(\Magento\Framework\App\ResourceConnection::DEFAULT_CONNECTION);
+        $connection = $this->resource->getConnection(ResourceConnection::DEFAULT_CONNECTION);
+
         $select = $connection->select()
-            ->from(
-                ['pu' => $productUpdatesTable],
-                ['product_id']
-            )
+            ->from(['pu' => $productUpdatesTable], ['product_id'])
             ->limit($this->syncLimit);
+
         $data = $connection->fetchAll($select);
 
-        if (is_array($data)) {
+        if (is_array($data) && count($data) > 0) {
             // Process the current batch
             foreach ($data as $productIdRow) {
                 $productId = $productIdRow['product_id'];
-                
-                // Delete the current product ID so we don't process it again
-                $connection->delete(
-                    $productUpdatesTable,
-                    ['product_id = ?' => $productId]
-                );
 
-                $this->printMessage("BulkUpdateProcessor - ProductId: $productId", $output);
+                try {
+                    // Delete the current product ID so we don't process it again
+                    // We do this FIRST so if it crashes, it doesn't get stuck in an infinite loop
+                    $connection->delete(
+                        $productUpdatesTable,
+                        ['product_id = ?' => $productId]
+                    );
 
-                // Default store view
-                $this->printMessage("BulkUpdateProcessor - XML default store", $output);
+                    $this->printMessage("BulkUpdateProcessor - ProductId: $productId", $output);
 
-                $data = $this->cuproducts->generateCuXmlForSingleProduct($productId, 0);
-                // Send to CU
-                $this->helper->updateProductData(0, $data);
+                    // Default store view
+                    $this->printMessage("BulkUpdateProcessor - XML default store", $output);
+                    $productData = $this->cuproducts->generateCuXmlForSingleProduct($productId, Store::DEFAULT_STORE_ID);
 
-                // Loop through store views
-                $websites = $this->storeManager->getWebsites();
-                foreach ($websites as $website) {
-                    $stores = $website->getStores();
-                    foreach ($stores as $storeView) {
-                        $storeId = $storeView->getData('store_id');
-                        $this->printMessage("BulkUpdateProcessor - storeId: $storeId", $output);
+                    // Send to CU
+                    $this->helper->updateProductData(Store::DEFAULT_STORE_ID, $productData);
 
-                        $data = $this->cuproducts->generateCuXmlForSingleProduct($productId, $storeId);
-                        $this->printMessage("BulkUpdateProcessor - XML generated", $output);
+                    // Loop through store views
+                    $websites = $this->storeManager->getWebsites();
+                    foreach ($websites as $website) {
+                        $stores = $website->getStores();
+                        foreach ($stores as $storeView) {
+                            $storeId = $storeView->getId();
+                            $this->printMessage("BulkUpdateProcessor - storeId: $storeId", $output);
 
-                        // Send to CU
-                        $this->helper->updateProductData($storeId, $data);
+                            $storeProductData = $this->cuproducts->generateCuXmlForSingleProduct($productId, $storeId);
+                            $this->printMessage("BulkUpdateProcessor - XML generated", $output);
+
+                            // Send to CU
+                            $this->helper->updateProductData($storeId, $storeProductData);
+                        }
                     }
+                } catch (\Exception $e) {
+                    $this->printMessage("BulkUpdateProcessor - Error processing ProductId $productId: " . $e->getMessage(), $output);
                 }
+
                 $currTime = time();
-                if ($currTime - $startTime > 5*60) {
-                    // Taking too long (more than 5 minutes), bail out
-                    $this->printMessage("BulkUpdateProcessor - Terminating early", $output);
+                // Taking too long (more than 5 minutes), bail out
+                if ($currTime - $startTime > 300) {
+                    $this->printMessage("BulkUpdateProcessor - Terminating early to prevent overlap", $output);
                     return;
                 }
             }
-        
+
             $this->printMessage("BulkUpdateProcessor - Complete", $output);
         }
     }
-    
+
     /**
-     * Logs an activity message and prints to the screen if being run as a
-     * command.
-     *
-     * @param type $str
-     * @param type $output
+     * Logs an activity message and prints to the screen if being run as a command.
      */
-    private function printMessage($str, $output)
+    private function printMessage(string $str, OutputInterface $output = null)
     {
         $this->helper->logInfo($str);
         if ($output) {
